[*] extract adapter class from main activity
